refactor: add `section` folder
