+++ src/components/index/Features.tsx
@@ -1,8 +1,8 @@
 import { memo } from "react";
 
 import { Card } from "./Card";
-import { SectionHeading } from "./SectionHeading";
-import { SectionLayout } from "./SectionLayout";
+import { SectionHeading } from "./section/SectionHeading";
+import { SectionLayout } from "./section/SectionLayout";
 
 const FEATURES_CARDS = [
   {
